+++ azure/services/dns/cnamerecords.go
@@ -65,8 +65,7 @@ func (s *Service) updateCnameRecords(ctx context.Context, currentRecordSets []*a
 
 func (s *Service) calculateMissingCnameRecords(logger logr.Logger, currentRecordSets []*armdns.RecordSet) []*armdns.RecordSet {
 
-	clusterZoneName := s.scope.ClusterDomain()
-	desiredRecords := desiredCnameRecords(clusterZoneName)
+	desiredRecords := desiredCnameRecords(s.scope.WildcardFQDN())
 
 	var recordsToCreate []*armdns.RecordSet
 
@@ -93,7 +92,7 @@ func (s *Service) calculateMissingCnameRecords(logger logr.Logger, currentRecord
 	return recordsToCreate
 }
 
-func desiredCnameRecords(clusterZoneName string) []*armdns.RecordSet {
+func desiredCnameRecords(wildcardTarget string) []*armdns.RecordSet {
 	return []*armdns.RecordSet{
 		{
 			Name: pointer.String("*"),
@@ -101,7 +100,7 @@ func desiredCnameRecords(clusterZoneName string) []*armdns.RecordSet {
 			Properties: &armdns.RecordSetProperties{
 				TTL: pointer.Int64(cnameRecordTTL),
 				CnameRecord: &armdns.CnameRecord{
-					Cname: pointer.String(fmt.Sprintf("ingress.%s", clusterZoneName)),
+					Cname: pointer.String(wildcardTarget),
 				},
 			},
 		},