+++ azure/services/privatedns/cnamerecords.go
@@ -66,8 +66,7 @@ func (s *Service) updateCnameRecords(ctx context.Context, currentRecordSets []*a
 
 func (s *Service) calculateMissingCnameRecords(logger logr.Logger, currentRecordSets []*armprivatedns.RecordSet) []*armprivatedns.RecordSet {
 
-	clusterZoneName := s.scope.ClusterDomain()
-	desiredRecords := desiredCnameRecords(clusterZoneName)
+	desiredRecords := desiredCnameRecords(s.scope.WildcardFQDN())
 
 	var recordsToCreate []*armprivatedns.RecordSet
 
@@ -94,7 +93,7 @@ func (s *Service) calculateMissingCnameRecords(logger logr.Logger, currentRecord
 	return recordsToCreate
 }
 
-func desiredCnameRecords(clusterZoneName string) []*armprivatedns.RecordSet {
+func desiredCnameRecords(wildcardTarget string) []*armprivatedns.RecordSet {
 	return []*armprivatedns.RecordSet{
 		{
 			Name: pointer.String("*"),
@@ -102,7 +101,7 @@ func desiredCnameRecords(clusterZoneName string) []*armprivatedns.RecordSet {
 			Properties: &armprivatedns.RecordSetProperties{
 				TTL: pointer.Int64(cnameRecordTTL),
 				CnameRecord: &armprivatedns.CnameRecord{
-					Cname: pointer.String(fmt.Sprintf("ingress.%s", clusterZoneName)),
+					Cname: pointer.String(wildcardTarget),
 				},
 			},
 		},